requirements.txt
Django>=5.1.3
djangorestframework>=3.15.2
django-storages>=1.14.4
boto3>=1.35.55
psycopg2-binary>=2.9.10
redis>=5.2.0
django-redis>=5.4.0
Pillow>=11.0.0
python-dotenv>=1.0.1
requests>=2.31.0
gunicorn>=21.2.0